fix(admin): added bp version on admin ui
